requirements.txt
ed25519==1.5
nats-py==2.9.0
nkeys==0.1.0
